+++ tableau_documents/__init__.py
@@ -1,6 +1,5 @@
 from tableau_connection import TableauConnection
 from tableau_datasource import TableauDatasource
-from tableau_datasource_generator import TableauDatasourceGenerator
 from tableau_document import TableauColumns, TableauDocument
 from tableau_file import TableauFile
 from tableau_workbook import TableauWorkbook
