+++ tableau_rest_api/__init__.py
@@ -1,2 +1,13 @@
-from tableau_rest_api_connection import TableauRestApiConnection, TableauRestApiConnection21, TableauRestApiConnection22, TableauRestApiConnection23, TableauRestApiConnection24, TableauRestApiConnection25, TableauRestApiConnection26, Sort, UrlFilter, ContentDeployer
+from tableau_rest_api_connection import TableauRestApiConnection
+from tableau_rest_api_connection_21 import TableauRestApiConnection21
+from tableau_rest_api_connection_22 import TableauRestApiConnection22
+from tableau_rest_api_connection_23 import TableauRestApiConnection23
+from tableau_rest_api_connection_24 import TableauRestApiConnection24
+from tableau_rest_api_connection_25 import TableauRestApiConnection25
+from tableau_rest_api_connection_26 import TableauRestApiConnection26
+from tableau_rest_api_connection_27 import TableauRestApiConnection27
+from tableau_rest_api_connection_28 import TableauRestApiConnection28
+from sort import Sort
+from url_filter import UrlFilter, UrlFilter23, UrlFilter24, UrlFilter25, UrlFilter26, UrlFilter27, UrlFilter28
+from content_deployer import ContentDeployer
 from published_content import Project20, Project21, Workbook, Datasource
